refactor: move sumcheck to an independent crate
