+++ sumcheck/src/layout/prover/prefix.rs
@@ -11,16 +11,16 @@ use p3_multilinear_util::point::Point;
 use p3_multilinear_util::poly::Poly;
 use p3_multilinear_util::split_eq::SplitEq;
 
-use crate::pcs::committer::writer::commit_base;
-use crate::sumcheck::lagrange::lagrange_weights_01inf_multi;
-use crate::sumcheck::layout::opening::Opening;
-use crate::sumcheck::layout::prover::Layout;
-use crate::sumcheck::layout::witness::{Table, TablePlacement};
-use crate::sumcheck::layout::{LayoutStrategy, ProverMultiClaim, ProverVirtualClaim, Witness};
-use crate::sumcheck::product_polynomial::ProductPolynomial;
-use crate::sumcheck::strategy::{SumcheckProver, VariableOrder};
-use crate::sumcheck::svo::{SvoPoint, calculate_accumulators_batch};
-use crate::sumcheck::{Claim, SumcheckData, extrapolate_01inf};
+use crate::commit::commit_base;
+use crate::lagrange::lagrange_weights_01inf_multi;
+use crate::layout::opening::Opening;
+use crate::layout::prover::Layout;
+use crate::layout::witness::{Table, TablePlacement};
+use crate::layout::{LayoutStrategy, ProverMultiClaim, ProverVirtualClaim, Witness};
+use crate::product_polynomial::ProductPolynomial;
+use crate::strategy::{SumcheckProver, VariableOrder};
+use crate::svo::{SvoPoint, calculate_accumulators_batch};
+use crate::{Claim, SumcheckData, extrapolate_01inf};
 
 /// Stacked-sumcheck prover with prefix-first variable binding.
 ///
