+++ whir/src/lib.rs
@@ -9,9 +9,9 @@
 
 extern crate alloc;
 
-pub mod constraints;
 pub mod fiat_shamir;
 pub mod parameters;
 pub mod pcs;
-pub mod sumcheck;
 pub(crate) mod utils;
+
+pub use p3_sumcheck::{self as sumcheck, constraints};
